+++ sources/Google.Solutions.Apis/Auth/Gaia/GaiaOidcSession.cs
@@ -1,0 +1,69 @@
+﻿//
+// Copyright 2023 Google LLC
+//
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+// 
+//   http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+
+using Google.Apis.Auth.OAuth2;
+using Google.Solutions.Common.Util;
+using System.Diagnostics;
+
+namespace Google.Solutions.Apis.Auth.Gaia
+{
+    internal class GaiaOidcSession : IGaiaOidcSession
+    {
+        private readonly UserCredential apiCredential;
+
+        public GaiaOidcSession(
+            IDeviceEnrollment deviceEnrollment,
+            UserCredential apiCredential,
+            IJsonWebToken idToken)
+        {
+            this.DeviceEnrollment = deviceEnrollment.ExpectNotNull(nameof(deviceEnrollment));
+            this.apiCredential = apiCredential.ExpectNotNull(nameof(apiCredential));
+            this.IdToken = idToken.ExpectNotNull(nameof(idToken));
+
+            Debug.Assert(idToken.Payload.Email != null);
+        }
+
+        public IJsonWebToken IdToken { get; }
+        public ICredential ApiCredential => this.apiCredential;
+        public IDeviceEnrollment DeviceEnrollment { get; }
+        public string Username => this.IdToken.Payload.Email;
+        public string Email => this.IdToken.Payload.Email;
+        public string HostedDomain => this.IdToken.Payload.HostedDomain;
+
+        public OidcOfflineCredential OfflineCredential
+        {
+            get
+            {
+                //
+                // Prefer fresh ID token if it's available, otherwise
+                // use old.
+                //
+                var idToken = string.IsNullOrEmpty(this.apiCredential.Token.IdToken)
+                    ? this.IdToken.ToString()
+                    : this.apiCredential.Token.IdToken;
+
+                return new OidcOfflineCredential(
+                    this.apiCredential.Token.RefreshToken,
+                    idToken);
+            }
+        }
+    }
+}
