+++ sources/Google.Solutions.Apis/Auth/Gaia/IGaiaOidcSession.cs
@@ -21,8 +21,19 @@
 
 namespace Google.Solutions.Apis.Auth.Gaia
 {
+    /// <summary>
+    /// A Google "1PI" OIDC session.
+    /// 
+    /// Sessions are subject to the 'Google Cloud Session Lenth' control.
+    /// </summary>
     internal interface IGaiaOidcSession : IOidcSession
     {
+        /// <summary>
+        /// ID token for the signed-in user.
+        /// Not null.
+        /// </summary>
+        IJsonWebToken IdToken { get; }
+
         /// <summary>
         /// Primary email address of user.
         /// </summary>
